Fixing Some UI Bugs
