+++ client/src/components/Form.jsx
@@ -46,6 +46,7 @@ function Form() {
             `${BASE_URL}?latitude=${lat}&longitude=${lng}`
           );
           const data = await res.json();
+
           console.log(data);
 
           if (!data.countryCode)
